# doc/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "WeFile API",
    "version": "0.0.1"
  },
  "basePath": "/api/v1",
  "paths": {
...
    "/user/group": {
      "get": {
        "description": "得到创建信息",
        "tags": [
          "Group"
        ],
        "summary": "GetGroup",
        "operationId": "GetGroup",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Cookie",
            "description": "存储 session id",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "GroupID",
            "description": "group id",
            "name": "group_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GroupResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/responses/ForbiddenResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "description": "创建 group",
        "tags": [
          "Group"
        ],
        "summary": "CreateGroup",
        "operationId": "CreateGroup",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Cookie",
            "description": "存储 session id",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "name",
                "password"
              ],
              "properties": {
                "name": {
                  "description": "group 名字",
                  "type": "string",
                  "maxLength": 64,
                  "minLength": 1,
                  "x-go-name": "Name"
                },
                "password": {
                  "description": "group 密码 用于检验 能否加入",
                  "type": "string",
                  "maxLength": 64,
                  "minLength": 8,
                  "x-go-name": "Password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GroupResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthorizedResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "description": "删除一个组，只允许创建者删除",
        "tags": [
          "Group"
        ],
        "summary": "DeleteGroup",
        "operationId": "DeleteGroup",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Cookie",
            "description": "存储 session id",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "GroupID",
            "description": "group id",
            "name": "group_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GroupResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthorizedResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/user/group/member_list": {
      "get": {
        "description": "得到组成员",
        "tags": [
          "Group"
        ],
        "summary": "GetGroupMemberList",
        "operationId": "GetGroupMemberList",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Cookie",
            "description": "存储 session id",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "GroupID",
            "description": "group id",
            "name": "group_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GroupMemberListResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          },
          "401": {
            "$ref": "#/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/responses/ForbiddenResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/user/sign_in": {
      "post": {
        "description": "用户登录",
        "tags": [
          "User"
        ],
        "summary": "SignIn",
        "operationId": "SignIn",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "description": "邮箱",
                  "type": "string",
                  "x-go-name": "Email"
                },
                "password": {
                  "description": "密码",
                  "type": "string",
                  "x-go-name": "Password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UserResponse"
          },
          "403": {
            "$ref": "#/responses/ForbiddenResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/user/sign_up": {
      "post": {
        "description": "用户注册",
        "tags": [
          "User"
        ],
        "summary": "SignUp",
        "operationId": "SignUp",
        "parameters": [
          {
            "x-go-name": "User",
            "description": "新用户",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UserResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          },
          "409": {
            "$ref": "#/responses/ConflictError"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Group": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "group_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "GroupID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "OwnerID"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/TensShinet/WeFile/service/base/handler"
    },
    "GroupUserInfo": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "group_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "GroupID"
        },
        "join_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "JoinAt"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "user_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/TensShinet/WeFile/service/base/handler"
    },
...
    "NewUser": {
      "type": "object",
      "required": [
        "name",
        "password",
        "email",
        "verify_code"
      ],
      "properties": {
        "email": {
          "description": "邮箱\n\n邮箱唯一",
          "type": "string",
          "x-go-name": "Email"
        },
        "name": {
          "description": "用户名",
          "type": "string",
          "maxLength": 64,
          "minLength": 1,
          "x-go-name": "Name"
        },
        "password": {
          "description": "密码",
          "type": "string",
          "maxLength": 64,
          "minLength": 8,
          "x-go-name": "Password"
        },
        "verify_code": {
          "description": "验证码\n\n邮件拿验证码",
          "type": "string",
          "x-go-name": "VerifyCode"
        }
      },
      "x-go-package": "github.com/TensShinet/WeFile/service/base/handler"
    },
...
    "User": {
      "type": "object",
      "properties": {
        "csrf_token": {
          "description": "csrf Token",
          "type": "string",
          "x-go-name": "CSRFToken"
        },
        "email": {
          "description": "邮箱",
          "type": "string",
          "x-go-name": "Email"
        },
        "name": {
          "description": "用户名",
          "type": "string",
          "x-go-name": "Name"
        },
        "profile": {
          "description": "个人简介",
          "type": "string",
          "x-go-name": "Profile"
        },
        "role_id": {
          "description": "角色 id",
          "type": "integer",
          "format": "int64",
          "x-go-name": "RoleID"
        },
        "role_name": {
          "description": "角色名",
          "type": "string",
          "x-go-name": "RoleName"
        },
        "user_id": {
          "description": "用户 id",
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/TensShinet/WeFile/service/base/handler"
    }
  },
  "responses": {
...
    "BadRequestResponse": {
      "description": "BadRequest",
      "headers": {
        "message": {
          "type": "string"
        }
      }
    },
    "ConflictError": {
      "description": "Conflict Error",
      "headers": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "ErrorResponse": {
      "description": "Sever Error",
      "headers": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "ForbiddenResponse": {
      "description": "Forbidden Error",
      "headers": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "GroupMemberListResponse": {
      "description": "GroupMemberListResponse",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/GroupUserInfo"
        }
      }
    },
    "GroupResponse": {
      "description": "GroupResponse",
      "schema": {
        "$ref": "#/definitions/Group"
      }
    },
...
    "UnauthorizedResponse": {
      "description": "Unauthorized Error",
      "headers": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "UserResponse": {
      "description": "User",
      "schema": {
        "$ref": "#/definitions/User"
      },
      "headers": {
        "cookie": {
          "type": "string"
        }
      }
    }
  }
}